Move get activities query to @/lib
