+++ src/app/takken/[slug]/api.ts
@@ -1,16 +1,9 @@
 import { generateApiQuery } from '@/lib/api';
-import { GROUP_PAGE_QUERY, ACTIVITIES_QUERY } from './query';
+import { GROUP_PAGE_QUERY } from './query';
 
 export const getGroupPage = (slug: string): Promise<any> => {
   return generateApiQuery({
     variables: { slug: slug },
     query: GROUP_PAGE_QUERY,
-  });
-};
-
-export const getActivities = (slug: string, date: string): Promise<any> => {
-  return generateApiQuery({
-    variables: { slug: slug, currDate: date },
-    query: ACTIVITIES_QUERY,
   });
 };
