+++ src/app/takken/[slug]/query.ts
@@ -73,24 +73,4 @@ const GROUP_PAGE_QUERY = gql`
   }
 `;
 
-const ACTIVITIES_QUERY = gql`
-  query getActivities($slug: String, $currDate: Date) {
-    activities(
-      filters: { group: { slug: { eq: $slug } }, endDate: { gte: $currDate } }
-      sort: "startDate:asc"
-    ) {
-      data {
-        attributes {
-          title
-          startDate
-          startTime
-          endDate
-          endTime
-          description
-        }
-      }
-    }
-  }
-`;
-
-export { GROUP_PAGE_QUERY, ACTIVITIES_QUERY };
+export { GROUP_PAGE_QUERY };
