+++ src/lib/api/activities/queries.ts
@@ -1,0 +1,23 @@
+import gql from 'graphql-tag';
+
+const GET_ACTIVITIES_QUERY = gql`
+  query getActivities($slug: String, $currDate: Date) {
+    activities(
+      filters: { group: { slug: { eq: $slug } }, endDate: { gte: $currDate } }
+      sort: "startDate:asc"
+    ) {
+      data {
+        attributes {
+          title
+          startDate
+          startTime
+          endDate
+          endTime
+          description
+        }
+      }
+    }
+  }
+`;
+
+export { GET_ACTIVITIES_QUERY };
